[aes] Use separate permutation for clearing second share of masked regs
Previously, both shares of masked registers (key, internal state) were
cleared with the same pseudo-random data. As a result, the sum of the two
shares was always 0 after clearing. This knowledge could potentially be
exploited for SCA.

Therefore, this commit adds another, separate permutation that is applied
to the output of the clearing PRNG for clearing the second share of masked
registers. This new permutation is exposed as a compile-time random
netlist constant.

Signed-off-by: Pirmin Vogel <vogelpi@lowrisc.org>