[rom_ctrl] A script to scramble a ROM image
This takes a 32-bit ELF file as input (something like the
boot_rom_sim_verilator.elf that we already create as part of the SW
build).

It pads the file up to the expected ROM size with pseudo-random data,
adds ECC checksum bits, extending to 40 bits in width and then
scrambles the result, writing out a VMEM file at the end.

This is not yet a complete solution because we don't calculate the
expected digest for the top 8 bytes (at the moment, they are all
zero). We'll come back to that once we have a proper model of the
interaction between rom_ctrl and the KMAC block.

As well as code to load a 32-bit ELF and write a 40-bit VMEM file,
this commit also has code to load VMEM files because the initial
version of this patch consumed e.g. boot_rom_sim_verilator.32.vmem.
The problem is that the build process has already zero-padded this
file internally, which we don't want, so we have to load
the (segmented) ELF instead. However, the VMEM loading code works and
might be useful soon, so I've left it in for now.

Signed-off-by: Rupert Swarbrick <rswarbrick@lowrisc.org>