[dvsim] Prevent command echo suppression
At some point in the past, I had refactored the way SW images were
handled. I had added `.ONESHELL:` to the `sw_build` target, assuming
that the `.ONESHELL` would only affect that target (cause all recipes of
a target to execute in the same shell). Turns out, its more of a global
setting - it affects ALL recipes of all targets.

This had an unintended consequence:
For all targets, the first recipe is an echo command printing the name
of that target, the echoing of which is suppressed with an `@`. Because the
rest of the recipes get invoked in the same shell, their echoing is also
suppressed. This makes issues hard to debug.

In this change, the `.ONESHELL` is removed, and all recipes of
`sw_build` are invoked one the same line with a trailing `\`. The first
command is `set -e` which is the equivalent of all commands chained with
`&&` (which is already what we want).

Signed-off-by: Srikrishna Iyer <sriyer@google.com>