[dv/alert_handler] Fix regression error
Two regression errors related to alert_handler:
1. If reset interrupt esc_cycle cnt, scb prediction might have one cycle
off due to reset is async, and we are only counting cycles at negedge
clock.
2. SCB is cycle accurate to predict esc_cnt, and currently scb only
support sending alert right when it finishes (not wait an extra cycle)
I will eventually try to support that in scb, but right now to pass
nightly regression, we will use this temp fix.

Finally I reduced the runtime for entropy_test. As there are 29 alerts
in top-level, it is hard to hit all corner cases by increase the
runtime. Need to fine tune and constraints.

Signed-off-by: Cindy Chen <chencindy@google.com>