[kmac] Fix app_config_t to work with Verilator
Verilator doesn't have proper support for unpacked structs. Since we
don't need one here, we can just declare app_config_t to be packed.

Also, it seems to have rather strange behaviour with int enums when
used as fields in a struct. I've reported this on the Verilator bug
tracker as issue #2855 but there's an easy workaround here, which is
probably what we'd want for synthesizable code anyway.

Signed-off-by: Rupert Swarbrick <rswarbrick@lowrisc.org>