[flash_ctrl] Narrow flash_phy_op_e to 2 bits
Using a literal enum value as an array index triggers a Verilator
check if the enum type isn't of the expected width. This was
triggering lots of warnings when defining the "reqs" variable in
flash_phy_core.sv. Since we don't actually need PhyOps as a member of
the enum, hoist it out into a separate parameter.

Signed-off-by: Rupert Swarbrick <rswarbrick@lowrisc.org>