[dv/otp_ctrl] add stress_all_test
This PR adds a few sequences to stress_all_test.
The chanllege for stress_all test is that otp_ctrl digest function
requires a reset to issue after it to make it work. Not following up
with a reset creates many corner cases on either RTL and DV side.
Coming PRs will add more tests to stress_all.

Signed-off-by: Cindy Chen <chencindy@google.com>