[sw, dif_kmac] Add implementation of SHA-3 (blocking only)
Add support for SHA-3 operations to the KMAC DIF. These operations
do not support returning partial results (i.e. they block until the
operation is complete). Support for non-blocking operations will be
added in a future PR.

Note that the squeeze operation will need to be modified when support
for XOF operations, such as SHAKE, are added. I have marked this with
'TODO(XOF):' for now.

This change only contains a smoke test for now. Unit tests will be
added at a later date.

Signed-off-by: Michael Munday <mike.munday@lowrisc.org>