[dvsim] Set `Deploy.job_name` more robustly
It is possible to have to sim_cfgs to have the same same, added to a
primary config supplied to DVSIm to run a large regression. For example:
SRAM_RET and SRAT_MAIN supply the Hjson with the same `name` attribute.
This causes a conflict in the `job_name` setting in the `Deploy`
class object. This PR fixes that by using `scratch_path` directory name
instead.

Signed-off-by: Srikrishna Iyer <sriyer@google.com>