[dvsim] Fix #6061
In the Scheduler code, if `poll()` is called but no jobs are actually
running, it is likely due to a SIGINT event. In that case, signal that
the status of the running jobs changed by returning True, so that the
status printer can print the final status one last time before exiting.

Signed-off-by: Srikrishna Iyer <sriyer@google.com>